+++ packages/spectacular/src/index.ts
@@ -16,8 +16,8 @@ export * from './lib/feature-testing/navigation/spectacular-feature-location';
 export * from './lib/feature-testing/navigation/spectacular-feature-router';
 
 // Pipe testing API
-export * from './lib/pipe-testing/pipe-harness/create-pipe-test-harness';
-export * from './lib/pipe-testing/pipe-harness/create-pipe-test-harness-options';
+export * from './lib/pipe-testing/pipe-harness/create-pipe-harness';
+export * from './lib/pipe-testing/pipe-harness/create-pipe-harness-options';
 export * from './lib/pipe-testing/pipe-harness/spectacular-pipe-harness';
 
 // Shared testing API
